降价

+++ src/main/java/cn/wowtw_backend/model/infoGroup/TalentsGearBase.java
@@ -22,5 +22,5 @@ public class TalentsGearBase {
 
     private Boolean isExtra;
 
-    private Integer sort;
+    private Double sort;
 }
